+++ PingUI/Views/AskToEnableUpdateCheckView.axaml.cs
@@ -1,0 +1,12 @@
+using Avalonia.ReactiveUI;
+using PingUI.ViewModels;
+
+namespace PingUI.Views;
+
+public partial class AskToEnableUpdateCheckView : ReactiveUserControl<AskToEnableUpdateCheckViewModel>
+{
+	public AskToEnableUpdateCheckView()
+	{
+		InitializeComponent();
+	}
+}